refactor: split app timer flow helpers
